refactor: split main.rs into separate modules and library crate (#26)
Adds some organization using modules instead of everything being in
main.rs to prepare for future development, and moves most of the logic
to a library crate that gets used by the main dapper binary so that
people could potentially use dapper as a dependency in other projects.
