+++ Cargo.toml
@@ -18,3 +18,4 @@ streaming-iterator = "0.1"
 walkdir = "2.5.0"
 rayon = "1.10.0"
 rusqlite = { version = "0.32.1", features = ["bundled"] }
+lazy_static = "1.5.0"
